# Add support for Streaming API