--- a/ZstdNet.Benchmarks/.gitignore
+++ b/ZstdNet.Benchmarks/.gitignore
@@ -1,0 +1,1 @@
+BenchmarkDotNet.Artifacts