--- a/ZstdNet.Benchmarks/CompressionBenchmarks.cs
+++ b/ZstdNet.Benchmarks/CompressionBenchmarks.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Running;
+
+namespace ZstdNet.Benchmarks
+{
+    [MemoryDiagnoser]
+    public class CompressionBenchmarks
+    {
+        const int TestSize = 1024;
+
+        byte[] UncompressedData = new byte[TestSize];
+        byte[] CompressedData;
+
+        byte[] Buffer = new byte[Compressor.GetCompressBound(TestSize)];
+
+        Compressor Compressor = new Compressor(new CompressionOptions(1));
+        Decompressor Decompressor = new Decompressor();
+
+        public CompressionBenchmarks()
+        {
+            var r = new Random(0);
+            r.NextBytes(UncompressedData);
+
+            CompressedData = Compressor.Wrap(UncompressedData);
+        }
+
+        [Benchmark]
+        public void Compress1KBRandom()
+        {
+            Compressor.Wrap(UncompressedData, Buffer, 0);
+        }
+
+        [Benchmark]
+        public void Decompress1KBRandom()
+        {
+            Decompressor.Unwrap(CompressedData, Buffer, 0);
+        }
+    }
+}