--- a/ZstdNet.Benchmarks/Main.cs
+++ b/ZstdNet.Benchmarks/Main.cs
@@ -1,0 +1,12 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace ZstdNet.Benchmarks
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            BenchmarkRunner.Run<CompressionBenchmarks>();
+        }
+    }
+}