--- a/ZstdNet.Benchmarks/ZstdNet.Benchmarks.csproj
+++ b/ZstdNet.Benchmarks/ZstdNet.Benchmarks.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.10.11" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ZstdNet\ZstdNet.csproj" />
+  </ItemGroup>
+
+</Project>