--- a/ZstdNet.sln
+++ b/ZstdNet.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.3
+VisualStudioVersion = 15.0.27130.2010
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZstdNet", "ZstdNet\ZstdNet.csproj", "{8ADBEB19-A508-471D-87A0-7443EE086E9B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZstdNet.Tests", "ZstdNet.Tests\ZstdNet.Tests.csproj", "{8DD3694E-9532-4659-AA9F-BB01CBB9341B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZstdNet.Tests", "ZstdNet.Tests\ZstdNet.Tests.csproj", "{8DD3694E-9532-4659-AA9F-BB01CBB9341B}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZstdNet.Benchmarks", "ZstdNet.Benchmarks\ZstdNet.Benchmarks.csproj", "{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,16 +33,28 @@ Global
 		{8ADBEB19-A508-471D-87A0-7443EE086E9B}.Release|x86.Build.0 = Release|Any CPU
 		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x64.ActiveCfg = Debug|x64
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x64.Build.0 = Debug|x64
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x86.ActiveCfg = Debug|x86
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x86.Build.0 = Debug|x86
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x64.Build.0 = Debug|Any CPU
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Debug|x86.Build.0 = Debug|Any CPU
 		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x64.ActiveCfg = Release|x64
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x64.Build.0 = Release|x64
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x86.ActiveCfg = Release|x86
-		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x86.Build.0 = Release|x86
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x64.ActiveCfg = Release|Any CPU
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x64.Build.0 = Release|Any CPU
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x86.ActiveCfg = Release|Any CPU
+		{8DD3694E-9532-4659-AA9F-BB01CBB9341B}.Release|x86.Build.0 = Release|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Debug|x64.Build.0 = Debug|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Debug|x86.Build.0 = Debug|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Release|x64.ActiveCfg = Release|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Release|x64.Build.0 = Release|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Release|x86.ActiveCfg = Release|Any CPU
+		{6ACD682F-6E1B-4C7E-B60E-66302AD9E2E3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE