--- a/ZstdNet/Compressor.cs
+++ b/ZstdNet/Compressor.cs
@@ -82,7 +82,7 @@ namespace ZstdNet
 			var dstCapacity = dst.Length - offset;
 			size_t dstSize;
 			using(var srcPtr = new ArraySegmentPtr(src))
-			using(var dstPtr = new ArraySegmentPtr(new ArraySegment<byte>(dst, offset, dstCapacity)))
+			using(var dstPtr = new ArraySegmentPtr(dst, offset, dstCapacity))
 			{
 				if(Options.Cdict == IntPtr.Zero)
 					dstSize = ExternMethods.ZSTD_compressCCtx(cctx, dstPtr, (size_t)dstCapacity, srcPtr, (size_t)src.Count, Options.CompressionLevel);