… ZstdNet/ExternMethods.cs