This should be in pretty decent shape. I've added some more basic tests that validate round tripping. 

I've also validated this on a ~50MB log file at various a couple of compression levels. 

Some rough testing indicates ~550MB/s Compression and ~875MB/s at level 1.

Let me know what you think.

fixes #11