@AlgorithmsAreCool 2018-01-28
Also included in this is a change to remove allocations from non-streaming compression and decompression.

Here are some timing figures from before and after commit 22056f7

**Before Changes**

|              Method |        Mean |     Error |    StdDev |  Gen 0 | Allocated |
|-------------------- |------------:|----------:|----------:|-------:|----------:|
|   Compress1KBRandom | 14,023.5 ns | 73.360 ns | 68.621 ns | 0.0153 |      80 B |
| Decompress1KBRandom |    254.0 ns |  5.116 ns |  6.652 ns | 0.0186 |      80 B |

**After Changes**

|              Method |        Mean |      Error |     StdDev | Allocated |
|-------------------- |------------:|-----------:|-----------:|----------:|
|   Compress1KBRandom | 13,982.4 ns | 32.6861 ns | 30.5745 ns |       0 B |
| Decompress1KBRandom |    232.1 ns |  0.9317 ns |  0.7780 ns |       0 B |