@Grantelbob 2019-03-31
yeah, seems like that... Sad :/

As a workaround i downloaded and build your streaming fork manually (https://github.com/AlgorithmsAreCool/ZstdNet/tree/Streaming), right after swapping out Zstd DLL for newest version. Works great until now :)

Thank you very much for your efford!