@AlgorithmsAreCool 2019-03-31
I'm very glad to hear it works for you. Maybe I should try to fork it permanently